lab2 finished

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,8 +3,8 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
